feat: issue-3 moved to a shared package
