+++ src/yaddd/shared/specification.py
